@@ -3,18 +3,22 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@babel/runtime": "^7.8.3",
     "axios": "^0.18.0",
     "grommet": "^2.7.0",
+    "moment": "^2.24.0",
     "node-sass": "^4.11.0",
     "query-string": "^6.2.0",
-    "react": "^16.7.0",
+    "react": "^16.12.0",
     "react-calendar": "^2.19.0",
-    "react-dom": "^16.7.0",
+    "react-dates": "^21.8.0",
+    "react-dom": "^16.12.0",
     "react-modal": "^3.8.2",
     "react-number-format": "^4.0.8",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
     "react-transition-group": "^2.5.3",
+    "react-with-direction": "^1.3.1",
     "styled-components": "^4.1.3"
   },
   "scripts": {
